+++ DatatypesAndExtensions/ChromaprintResult.cs
@@ -1,0 +1,8 @@
+namespace Ass_Pain
+{
+    public class ChromaprintResult
+    {
+        public float duration;
+        public string fingerprint;
+    }
+}
