+++ MainActivity.cs
@@ -314,6 +314,7 @@ namespace Ass_Pain
                 {
                     Console.WriteLine(e);
                 }*/
+                Downloader.GetMusicBrainzIDFromFingerprint(stateHandler.Songs.Search("dark hour").First().Path);
             }).Start();
             stateHandler.SetView(this);
             receiver = new MyBroadcastReceiver(this);
